chore: update expo-router example with ShareIntentProvider
