+++ example/expo-router/app/[...unmatched].tsx
@@ -1,0 +1,4 @@
+// for iOS deeplink redirect to this unmatched page
+// catched link: exposhareintentexample:///dataUrl=exposhareintentexampleShareKey
+import Home from "./index";
+export default Home;
